todo-app/todoui-flask/requirements.txt
Faker==28.4.1
flask==3.0.3

opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-sdk==1.26.0
requests==2.32.3
